+++ src/Application/Resources.hpp
@@ -10,7 +10,7 @@
 
 #include "Utils/FontLoader.hpp"
 #include "Utils/TextureLoader.hpp"
-#include "Graphics/Texture2d.hpp"
+#include "Graphics/Texture2D.hpp"
 #include "Graphics/Font.hpp"
 #include "Graphics/Shader.hpp"
 
@@ -18,7 +18,7 @@ using std::map;
 using std::string;
 
 class Font;
-class Texture2d;
+class Texture2D;
 class Shader;
 
 class Resources {
@@ -28,11 +28,11 @@ class Resources {
     ~Resources();
   
     void AddFont(const string& _name, Font* _font);
-    void AddTexture(const string& _name, Texture2d* _texture);
+    void AddTexture(const string& _name, Texture2D* _texture);
     void AddShader(const string& _name, Shader* _shader);
   
     Font* GetFont(const string& _name);
-    Texture2d* GetTexture(const string& _name);
+    Texture2D* GetTexture(const string& _name);
     Shader* GetShader(const string& _name);  
     
     static Resources* Get();
@@ -40,7 +40,7 @@ class Resources {
     static Font* GetFont();
   private:
     map<string, Font*> m_fonts;
-    map<string, Texture2d*> m_textures;
+    map<string, Texture2D*> m_textures;
     map<string, Shader*> m_shaders;
     static Resources* m_instance;
 };