+++ src/Entities/Entity.cpp
@@ -6,6 +6,6 @@
 #include "utils/GLDebug.hpp"
 
 Entity::
-Entity(Texture2d* _tex, const glm::vec3& _loc,
+Entity(Texture2D* _tex, const glm::vec3& _loc,
 		const glm::vec2& _size, float _rotation, Entity_Type _type):
-		Renderable(_tex, _loc, _size, _rotation), m_type(_type) {}
+		Renderable2D(_tex, _loc, _size, _rotation), m_type(_type) {}