+++ src/Graphics/Renderer2D.hpp
@@ -8,23 +8,33 @@
 
 
 #include "Application/Resources.hpp"
+#include "Camera/Camera.hpp"
+#include "gl/VertexArray.hpp"
+#include "gl/VertexBuffer.hpp"
+#include "gl/IndexBuffer.hpp"
 #include "Font.hpp"
 #include "Utils/MatrixStack.h"
-#include "Renderable.hpp"
+#include "Renderable2D.hpp"
 #include "Shader.hpp"
 
 #define MAX_TEXTURE_SIZE 60000
 #define MAX_INDICE_SIZE MAX_TEXTURE_SIZE * 6
+#define MAX_BUFFER_SIZE (MAX_TEXTURE_SIZE * 4 * sizeof(VertexData))
 
 class Renderer {
+  
   public:
+
     Renderer(const glm::ivec2& _bufferSize);
     void Init();
 
-    void Submit(Renderable* _entity);
+    void Submit(Renderable2D* _entity);
+    void Begin();
     void Present();
+    void End();
 
     void DrawString(const std::string& _text, float _x, float _y, const Font& _font = *Resources::GetFont(), unsigned int _color = 0xffffffff);
+
     void DrawLine(float _x0, float _y0, float _x1, float _y1, unsigned int _color = 0xffffffff, float _thickness = 0.02);
     void DrawLine(const glm::vec2& _start, const glm::vec2& _end, unsigned int _color = 0xffffffff, float _thickness = 0.02);
     
@@ -41,20 +51,21 @@ class Renderer {
 
     void SetScreenSize(const glm::ivec2& _bufferSize);
 
+    void SetCamera(Camera* _camera);
   private:
-    float SubmitTexture(Texture2d* _texture);
-    //std::vector<Entity*> m_entities;
-    std::vector<Texture2d*> m_textures;
-    MatrixStack* m_matrixStack;
-    Shader* m_textureShader;
-    Shader* m_textShader;
-    GLuint m_vertexArray;
-    GLuint m_indexBuffer;
-    VertexData* m_vertexData;
-    VertexData* m_vertexHead;
-    size_t m_indexSize;
-    glm::ivec2 m_bufferSize;
-    // Camera m_camera;
+
+    float SubmitTexture(Texture2D* _texture);
+
+    std::vector<Texture2D*> m_textures; ///< List of currently rendered textures.
+    MatrixStack* m_matrixStack;         ///< Rendering matrix stack.
+    Shader* m_textureShader;            ///< Shader program for rendering textures.
+    Shader* m_textShader;               ///< Shader program fro rendering text.
+    VertexArray* m_vertexArray;         ///< Vertex Array object for rendering.
+    IndexBuffer* m_indexBuffer;         ///< Single buffer for storing indices.
+    VertexData* m_vertexData;           ///< Buffer for storing vertex buffer.
+    size_t m_indexSize;                 ///< Number of indices that should be rendered.
+    glm::ivec2 m_bufferSize;            ///< The size of the rendering buffer.
+    Camera* m_camera;                   ///< The scenes camera.
 };