+++ src/Graphics/Texture2d.cpp
@@ -1,13 +1,13 @@
-#include "Texture2d.hpp"
+#include "Texture2D.hpp"
 #include "utils/GLDebug.hpp"
-unsigned int Texture2d::m_nextId = 0;
+unsigned int Texture2D::m_nextId = 0;
 
-Texture2d::
-Texture2d(GLuint _gltex, int _width, int _height, unsigned int* _data): m_uid(m_nextId++), 
+Texture2D::
+Texture2D(GLuint _gltex, int _width, int _height, unsigned int* _data): m_uid(m_nextId++),
     m_texture(_gltex), m_width(_width), m_height(_height), m_loaded(_data != nullptr), m_unit(0) {}
 
-Texture2d::
-Texture2d(const Texture2d& _tex) {
+Texture2D::
+Texture2D(const Texture2D& _tex) {
     m_uid           = _tex.m_uid;
     m_texture       = _tex.m_texture;
     m_width         = _tex.m_width;
@@ -16,9 +16,9 @@ Texture2d(const Texture2d& _tex) {
     m_loaded        = _tex.m_loaded;
 }
 
-Texture2d
-Texture2d::
-operator= (const Texture2d& _tex) {
+Texture2D
+Texture2D::
+operator= (const Texture2D& _tex) {
     m_uid           = _tex.m_uid;
     m_texture       = _tex.m_texture;
     m_width         = _tex.m_width;
@@ -28,17 +28,18 @@ operator= (const Texture2d& _tex) {
 }
 
 void
-Texture2d::
-Bind(GLuint _handle) const {
+Texture2D::
+Bind() const {
     glActiveTexture(GL_TEXTURE0 + m_unit);
     glBindTexture(GL_TEXTURE_2D, m_texture);
-    glUniform1i(_handle, m_unit);
+    glCheckError();
 }
 
 
 void
-Texture2d::
+Texture2D::
 Unbind() const {
     glActiveTexture(GL_TEXTURE0 + m_unit);
     glBindTexture(GL_TEXTURE_2D, 0);
+    glCheckError();
 }