+++ changelog.md
@@ -26,7 +26,7 @@
 * Text Rendering works for the most part.
   * FTLoader: Partially works
   * Font: Works less than FTLoader.
-* Created a texture2d class.
+* Created a Texture2D class.
 
 ## 17/02/2017: Andrew
 * Running OpenGl 3.3 (can run 4.1).